+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Payment processing loop is continued if exchange rate for particular period is not available
+### Fixed
+- Do not process payments older than 2 days (by default)
 
 ## [2.4.12] - 2023-03-31
 ### Added